implements architecture modular android
