+++ app/src/main/AndroidManifest.xml
@@ -2,6 +2,9 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools">
 
+    <uses-permission android:name="android.permission.INTERNET" />
+    <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
+
     <application
         android:name=".ui.NewsletterApplication"
         android:allowBackup="true"
@@ -14,7 +17,7 @@
         android:theme="@style/Theme.MaterialComponents.Bridge"
         tools:targetApi="31">
         <activity
-            android:name=".ui.activity.MainActivity"
+            android:name=".ui.screen.home.MainActivity"
             android:exported="true"
             android:theme="@style/Theme.MaterialComponents.Bridge">
             <intent-filter>
@@ -24,8 +27,5 @@
             </intent-filter>
         </activity>
     </application>
-
-    <uses-permission android:name="android.permission.INTERNET" />
-    <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
 
 </manifest>
