+++ app/src/main/java/com/jd/newsletter/ui/NewsletterApplication.kt
@@ -4,5 +4,4 @@ import android.app.Application
 import dagger.hilt.android.HiltAndroidApp
 
 @HiltAndroidApp
-class NewsletterApplication : Application() {
-}
+class NewsletterApplication : Application()
