--- app/src/main/java/com/jd/newsletter/ui/activity/adapters/LoadingStateAdapter.kt
@@ -1,33 +1,0 @@
-package com.jd.newsletter.ui.activity.adapters
-
-import android.view.LayoutInflater
-import android.view.ViewGroup
-import androidx.core.view.isVisible
-import androidx.paging.LoadState
-import androidx.paging.LoadStateAdapter
-import androidx.recyclerview.widget.RecyclerView
-import com.jd.newsletter.databinding.LoadingItemBinding
-
-class LoadingStateAdapter : LoadStateAdapter<LoadingStateAdapter.LoadingStateViewHolder>() {
-
-    override fun onCreateViewHolder(
-        parent: ViewGroup,
-        loadState: LoadState
-    ): LoadingStateViewHolder {
-        return LoadingStateViewHolder(
-            LoadingItemBinding.inflate(
-                LayoutInflater.from(parent.context),
-                parent,
-                false
-            )
-        )
-    }
-
-    override fun onBindViewHolder(holder: LoadingStateViewHolder, loadState: LoadState) {
-        holder.binding.progressBar.isVisible = loadState is LoadState.Loading
-    }
-
-    inner class LoadingStateViewHolder(val binding: LoadingItemBinding) :
-        RecyclerView.ViewHolder(binding.root)
-
-}
