--- app/src/main/java/com/jd/newsletter/ui/data/model/NewsModelResponse.kt
@@ -1,12 +1,0 @@
-package com.jd.newsletter.ui.data.model
-
-data class NewsModelResponse(
-    val count: Long,
-    val page: Long,
-    val totalPages: Long,
-    val nextPage: Long,
-    val previousPage: Long,
-    val showingFrom: Long,
-    val showingTo: Long,
-    val items: List<NewsModel>
-)
