--- app/src/main/java/com/jd/newsletter/ui/repository/NewsletterRepository.kt
@@ -1,9 +1,0 @@
-package com.jd.newsletter.ui.repository
-
-import androidx.paging.PagingData
-import com.jd.newsletter.ui.data.model.NewsModel
-import io.reactivex.Flowable
-
-interface NewsletterRepository {
-    fun getNewsLetter(): Flowable<PagingData<NewsModel>>
-}
