--- app/src/main/java/com/jd/newsletter/ui/useCase/GetNewsLetterUseCase.kt
@@ -1,9 +1,0 @@
-package com.jd.newsletter.ui.useCase
-
-import com.jd.newsletter.ui.repository.NewsletterRepository
-
-class GetNewsLetterUseCase(
-    private val newsLetterRepository: NewsletterRepository
-) {
-    operator fun invoke() = newsLetterRepository.getNewsLetter()
-}
