--- app/src/main/java/com/jd/newsletter/ui/useCase/UseCaseNewsletter.kt
@@ -1,5 +1,0 @@
-package com.jd.newsletter.ui.useCase
-
-data class UseCaseNewsletter(
-    val getNewsletter: GetNewsLetterUseCase,
-)
