+++ data/src/main/java/com/jd/data/repository/dataSource/NewsletterRemoteDataSource.kt
@@ -1,0 +1,9 @@
+package com.jd.data.repository.dataSource
+
+import androidx.paging.PagingData
+import com.jd.domain.model.NewsModel
+import io.reactivex.Flowable
+
+interface NewsletterRemoteDataSource {
+    fun getNewsLetter(): Flowable<PagingData<NewsModel>>
+}
