+++ domain/build.gradle.kts
@@ -1,12 +1,10 @@
 plugins {
     alias(libs.plugins.android.library)
     alias(libs.plugins.jetbrains.kotlin.android)
-    alias(libs.plugins.org.jetbrains.kotlin.kapt)
-    alias(libs.plugins.dagger.hilt)
 }
 
 android {
-    namespace = "com.jd.data"
+    namespace = "com.jd.domain"
     compileSdk = 34
 
     defaultConfig {
@@ -36,25 +34,16 @@ android {
 
 dependencies {
 
-    implementation(project(":domain"))
-
     implementation(libs.androidx.core.ktx)
     implementation(libs.androidx.appcompat)
     implementation(libs.google.material)
-    testImplementation(libs.junit)
-    androidTestImplementation(libs.androidx.junit)
-    androidTestImplementation(libs.androidx.espresso.core)
-
+    implementation(libs.converter.gson)
     implementation(libs.rx.java)
     implementation(libs.rx.android)
-
     implementation (libs.androidx.paging.runtime.ktx)
     implementation (libs.androidx.paging.rxjava2.ktx)
 
-    implementation(libs.retrofit)
-    implementation(libs.dagger.hilt)
-    implementation(libs.hilt.android.v251)
-    kapt(libs.hilt.android.compiler)
-    annotationProcessor(libs.hilt.android.v251)
-
+    testImplementation(libs.junit)
+    androidTestImplementation(libs.androidx.junit)
+    androidTestImplementation(libs.androidx.espresso.core)
 }
