+++ domain/src/main/java/com/jd/domain/repository/NewsletterRepository.kt
@@ -1,0 +1,9 @@
+package com.jd.domain.repository
+
+import androidx.paging.PagingData
+import com.jd.domain.model.NewsModel
+import io.reactivex.Flowable
+
+interface NewsletterRepository {
+    fun getNewsLetter(): Flowable<PagingData<NewsModel>>
+}
