+++ domain/src/main/java/com/jd/domain/useCase/GetNewsLetterUseCase.kt
@@ -1,0 +1,7 @@
+package com.jd.domain.useCase
+
+import com.jd.domain.repository.NewsletterRepository
+
+class GetNewsLetterUseCase(private val newsLetterRepository: NewsletterRepository) {
+    operator fun invoke() = newsLetterRepository.getNewsLetter()
+}
