+++ domain/src/main/java/com/jd/domain/useCase/UseCaseNewsletter.kt
@@ -1,0 +1,5 @@
+package com.jd.domain.useCase
+
+data class UseCaseNewsletter(
+    val getNewsletter: GetNewsLetterUseCase,
+)
