+++ gradle/libs.versions.toml
@@ -23,8 +23,8 @@ googleMaterial = "1.12.0"
 org-jetbrains-kotlin-android = "1.9.0"
 
 [libraries]
-adapter-rxjava2 = { module = "com.squareup.retrofit2:adapter-rxjava2", version.ref = "adapterRxjava2" }
 androidx-core-ktx = { group = "androidx.core", name = "core-ktx", version.ref = "coreKtx" }
+adapter-rxjava2 = { module = "com.squareup.retrofit2:adapter-rxjava2", version.ref = "adapterRxjava2" }
 androidx-paging-runtime-ktx = { module = "androidx.paging:paging-runtime-ktx", version.ref = "pagingRuntimeKtx" }
 androidx-paging-rxjava2-ktx = { module = "androidx.paging:paging-rxjava2-ktx", version.ref = "pagingRuntimeKtx" }
 hilt-android-v251 = { module = "com.google.dagger:hilt-android", version.ref = "googleHiltAndroid" }
